Move `junit-formatter.go` to its own package for use as a library.
This lets JUnit xml generation be integrated into larger tools
instead of having to run the `go-junit-report` binary as a subprocess.
