+++ formatter/formatter.go
@@ -1,4 +1,4 @@
-package main
+package formatter
 
 import (
 	"bufio"
